renovate: try to see if we can accurately group our various deps
This is a learning exercise which aims to educate and enable:

- Grouping of dependency updates
- Auto-merging of pre-1.0 _patch_ updates (which are usually _not breaking)
  but NOT auto-merge pre-1.0 minor (which are more than usually breaking).

Yes, I know this is not per Semver for pre-1.0 (which suggests that
_everything_ is experimental and no bump, patch or minor should be trusted),
but so much of the Rust ecosystem is pre-1.0.

Administrative note: I have not opened this as a PR because honestly it's
just very difficult to test these things without landing them and seeing
what Renovate does.  On the plus side, nothing is set to auto-merge after
this PR, so reverting this commit should just make Renovate undo whatever it
does.

I'm also going to claim that "I know what I'm doing".  (and then probably
immediately electrocute myself)