Revert "Add debug information to release profile"
This reverts commit c04976e2bba6b33a74b99a226fc7737aee758f79.

Reverting for now because it is causing issue on my machine and I'm not
sure it will work either on release workflow on the CI.

[0] [13:46:19] ~/r/router main > ./target/release/router
Failed to execute process './target/release/router'. Reason:
exec: unknown error (errno was 8)
The file './target/release/router' is marked as an executable but could not be run by the operating system.