fix: indent and commas
While formatting .graphql files, my linter added commas and newlines around. Although this is allowed in the spec, and the parser handles it correctly, we weren't able to handle it when gathering endpoint names and urls.

This commits fixes this, and adds various indents and commas in the routing_urls test.