Warm the query planner cache with hot queries
If we use an LRU cache (crate: lru) for our query plans, then, whenever
we update queries or configuration, we can pick the top current query
keys and use those to pre-populate a new cache.

Most of this change is just expressing that simple idea.

There are yet questions to be resolved:
 - How large should the cache be and should it be configurable?
 - How many hot keys should we preserve on re-configuration?
 - How best to expose this functionality without affecting good
   code structure too much?

Hence this is a draft PR for early review.