Implement more sophisticated query plan cache delegation
This version avoids the twin evils of:
 - locking cache for entire duration of delegated get
 - OR doing multiple delegated get for single query

It's more complex, so maybe not desirable on those grounds and does
introduce new failure paths and a new crate (bus). If we like the
approach, I could probably get rid of the bus crate and do this using
condvar or a different spmc implementation.