Remove bus crate and ensure all tests are passing
I wasn't happy with the bus crate, so this change removes it and makes
use of the tokio broadcast channel. Maybe not as fast, but well
supported and always good to use less crates.

All the mock tests are passing now, but I'm not really happy with the
changes I made. Might need more detailed examination.