Draft of proposed changes for cache unification
I don't really love either approach to unifying:
 - delegate callback
 - function parameter

Ideally we could store a callback in the CachingMap struct, but
lifetimes and async restrictions mean I can't make that work at the
moment.

This draft PR is to discuss the pros/cons of the two approaches and see
if anyone has suggestions for ways to improve.