Cache unification
Unify the two cache implementations into a single implementation.

After spending a bit of time exploring different ideas, I've settled
on the idea proposed by Geoffrey to provide a value resolver for use
by the cache map.

This approach is cleanest from the point of view of client interactions,
although it does require clients to implement the CacheResolver trait.

Although CacheResolver is using dynamic dispatch, it's only called for
once per cache miss (which is expected to be slow by definition), so
I think that's an acceptable performance tradeoff.