fix: Reinstall node in the circleCI release pipeline.
During the various iterations on the build pipeline, I ended up removing nodejs, and putting it back eventually. I unfortunately forgot to put it back in the release pipeline, which causes it to fail.

We have followup work in federation that shall remove the nodejs requirement eventually, but we're not there yet.