find an alternate way to merge response data
by delaying the merge of the subgraph response data into the global
response, we can avoid the Mutex over the response.
In particular, merging the responses from a parellel query plan
execution can be done by wrapping the sub queries in a FuturesUnordered,
that will make a stream of responses, that can then all be merged
together one by one.

Since subgraph responses do not have direct access to the global
responses, merging in fetch_node had to be changed, because it would try
to put JSON nodes in an empty JSON Value.
This is fixed by modifying the generation of representations data to
create a list of paths. That way, for each entity of the subgraph
response, we know the exact path where we want to insert it in the JSON
response. As a side effect of that work, this will fix #248