Fix: router-bridge: Do not panic when serializing or deserializing to interact with deno-core
The router bridge used to panic under certain circumstances, one of them being when deserialization fails.

Gary has unfortunately found that the hard way, when querying an (unknown) introspection query against the router.

It turns out introspection queries return an empty query plan, because there is effectively not much to query downstream. We however thought the root node of a query plan was mandatory, which ended up in a panic on the plan() call.

This PR follows up on the federation counterpart: https://github.com/apollographql/federation/pull/1321

Federation PR TLDR: We now have one more layer of Result, indicating whether an uncaught deno runtime error (like a deserialization error for example)

There's a couple of things going on:
- Update the router bridge dependency and rework the result types
- Make PlannerResult::node an Option<PlanNode>
- Add regression tests
- Identify the queries Gary ran (it was using npx diagnose-endpoint) and add the query strings to the well_known_queries NaiveIntrospection relies on, so they hit the cache.