experiment with streaming JSON serialization
instead of serializing the JSON response to one buffer, that needs to be
reallocated multiple times if it is not large enough, we could serialize
to a serie of small buffers that are sent as soon as they are full.

Apparently it adds some overhead in the smaller percentiles, but overall
results in a more stable latency distribution, and clients can start
parsing and processing data earlier.

I suspect this can have a big impact on larger response sizes