Add support for configurable collectors
Configuration now has support for a studio usage section where
an agent may be specified to be external or internal and an address for
an agent is specified. (XXX maybe ignore if not external...)

I've cleaned up the creation of agent to respect these changes and tried
to improve the use of tokio runtime, although that's not completely as
I'd like it yet.