try (and fail) to materialize the future returned by ApolloRouterService
this is done a lot in tower's code: a Service implementation has an
associated Future type that will perform the work. Probably to avoid
boxing everything.

Currently, trying this future fails with the error:
error[E0515]: cannot return value referencing local variable `router`
   --> apollo-router/src/apollo_router.rs:213:9
    |
212 |           let future = router.prepare_query(r);
    |                        ----------------------- `router` is
borrowed here
213 | /         ApolloRouterServiceFuture {
214 | |             state: State::Prepare { req: r2, future },
215 | |         }
    | |_________^ returns a value referencing data owned by the current
function

this may be fixed if we replace the Router instance with a Service
instance