support Service instances in the http service registry
This replaces the Fetcher trait with a SubgraphService trait, that
is implemented on Send+Sync+Service<Request, Response=Response,
Error=FetchError>+'static. That trait is not strictly necessary but I've
found it helps in keeping the function signature manageable, and since
we'll have various combinations of Service implementations, it feels
useful to have a canonical name.

The Fetcher trait and the previous implementation of ServiceRegistry
work on &self so they can be shared between threads easily, but we won't
be able to do that going forward, because some layers or extensions
might need mutable state (example: rate limiting).
That's why the Service::call method uses &mut self (the alternative would be
to use interior mutability). That's linked to another point about the
Service trait: instances are not meant to be shared between connections,
the basic use case is creating an owned instance for each connection, and it
will be used on each request (or a new one can be created for each
request).

So we need the ServiceRegistry to return instances of Service on which
we can call methods with mutable references. A way to do that is making
the service clonable (any non clonable service can be wrapped with Buffer
instead: https://docs.rs/tower/latest/tower/buffer/index.html ).

The service registry has to store instances with erased types because we
might have different kinds of fetchers depending on the subgraph. They
are cloned when requesting the service. We call the ServiceExt::oneshot`
method instead of call() because it consumes the Service instance
directly and avoids issues where the compiler assumes the Service
refernce could be used in the returned Future