Store JSON strings in a Bytes instance (#284)
One of the router's main tasks is to deserialize, filter, aggregate and serialize arbitrary JSON objects, for which the schema is not statically known, so we have to use serde_json::Value to represent them. This replaces Value with an equivalent type from the serde_json_bytes crate, that replaces string values and object keys with wrappers around bytes::Bytes. That way, we can have them refer to the buffer that contained the JSON data, instead of allocating and copying for every string. Especially considering that most of them will never be used directly by the router.

We will also get more benefits when implementing query deduplication and caching, as a single buffer will be usable by multiple responses at the same time.

Co-authored-by: Cecile Tonglet <cecile.tonglet@cecton.com>