configs which contain jaeger config will hang the router when updated
If a configuration is updated which contains jaeger configuration, then
at some point a synchronous call to:
  opentelemetry::global::set_tracer_provider()
will arise. Even though this call is made from a synchronous function,
the root context is asynchronous and it seems to cause the runtime to
mis-behave: i.e.: hang...

If we make the call from a newly spawned thread (no performance concerns
here because re-configuration is not hot code) then the problem goes
away.