Add support for studio integration into router configuration
So far in the prototype I have been hard-coding the graph reference
into the code and setting the key in the environment.

This updates the router configuration to add a new (optional) section
for graph details. If not present, then the agent is not configured.

The relayer now has a more complex mapping of tpq details, since each
submission of statistics must provide details about which graph they are
relevant for.

NB: For performance reasons, it may later turn out to be a good idea to
have different versions of the collector. A Single Graph version for
internal consumption which avoids passing in the graph details for every
submission and just configures a single set of details.

Documentation still has to be updated to reflect these changes.