add support for retries and triggering transfers
The server will still attempt to relay statistics every 5 seconds. In
addition, if we detect there are too many entries pending (10 for
testing, but increase later) we will try to trigger an early transfer.

The new triggering mechanism will re-try the trigger with a backoff if
notification fails. The send of data to the ingress is now also re-tried
with a backoff on failure.

If the send of data to the ingress ultimately fails, then the
accumulated reports at that point will be thrown away.

These measures should in combination control memory usage within the
server and maximise our ability to transfer data reliably.