remove the Extensions type
the Extensions struxture is not adapted to sharing data with dynamic
plugins, and introduces high coupling between extensions through the
type ids.
This replaces Extensions with a JSON map, so that data can be accessed
by a key, and transferred easily to dynamic plugins.
The RwLock in Context was also removed: we do not need synchronized
access on the context until we get to query plan execution.
The Extensions struct can be reevaluated after we get more experience
writing plugins