add support for apollo telemetry when OTLP tracing enabled
This was difficult to figure out. Turns out that OTLP has it's own set
of exporters that it's happy with (gRPC, HTTP etc...) and the OTLP
pipeline expects exporters that comply with that scheme...

So, rather than add in an exporter, like we did for Jaeger support,
it's simpler to add another layer, like we do if neither OTLP/Jaeger
tracing is enabled.