Give each service its own set of request and response types. (#416)
* Give each service its own set of request and response types.

The rationale behind this is that with the types as they were it was impossible to send a response payload of anything other than a graphql response. This would potentially cause issues for users wanting to send back specific error codes or messages for something such as auth.

The second reason is that there was some confusion around the types within the developers because sometimes a response was sent on the request side of a service and vice versa. This now is resolved with each type being used for one purpose only.

Before:
* RouterRequest
* PlannedRequest
* SubgraphRequest
* RouterResponse

Now:
* RouterRequest
* RouterResponse
* QueryPlannerRequest
* QueryPlannerResponse
* ExecutionRequest
* ExecutionResponse
* SubgraphRequest
* SubgraphResponse

* Add `map` for `Request` for completeness.

* Simplify match expression

* debug span -> trace span

* Remove unused field.

Co-authored-by: bryn <bryn@apollographql.com>