Dogfood layer config. Implement schema generation for layers.
This slightly differs from the format as described in the discussion thread.
kind has been converted to a key.

Actual user experience may not be great. A layer should probably do only one thig, so headers having a list of operations may be confusing.

Plugin factory should be modified to use the same sort of pattern as layer factory.