filter paths with values
in the query plan's fetch nodes, for entities queries, we must create a list of
variables and paths at which they must be inserted in the final
response. In the previous version of the code, we were creating both
lists, then filtering the values depending on the query.

In the case where the entities are different types implementing the same
interface, provided by different subgraphs, we could have the
following query:

query {
  nodes(ids: ["UHJvZHVjdDox", "UmV2aWV3OjI="]) {
    id
    ... on Product {
      name
    }
    ... on Review {
      body
    }
  }
}

We would get the response [{ "__typename": "Product", "id":
"UHJvZHVjdDox"}, {"__typename": "Review", "id": "UmV2aWV3OjI="}]

This would make a subquery to the products subgraph, and one to the
review subgraph.

The failing code was generating for the reviews query a list of paths
for the indexes 0 and 1 in that array, and both values, then
filtering the list to only keep {"__typename": "Review", "id":
"UmV2aWV3OjI="}

When returning the response, it would be inserted at the index 0, in the
product entity.

This commit fixes variable creation to filter both path and value at the
same time