Check the type condition on fragments
Fragments must specify which type they apply on. With the previous code,
all fragments were applied on all objects indiscriminately, which
resulted in fields appearing where they should not

As an example, with the query:

```graphql
query {
  nodes(ids: ["1", "2"]) {
    id
    ... on Product {
      name
    }
    ... on Review {
      body
      __typename
    }
  }
}
```

we would get this response:

```json
{
  "data": {
    "nodes": [
      {
        "id": "1",
        "name": "My little product",
        "__typename": "Product"
      },
      {
        "id": "2",
        "body": "It's pretty good!",
        "__typename": "Review"
      }
    ]
  }
}
```

Here __typename must only be displayed on the Review object.

This commit parses the type condition along with the selection set in
fragments, and checks that condition. It assumes that __typename will
always be available