make an enum for jaeger agent/collector configuration
both options are incompatible, so we must set up an enum. It is an
option, the default case is in UDP agent mode with a default address.
The collector mode requires a HTTP address, and the username and
password options only make sense in collector mode.
Username and password are kept in the main jaeger configuration struct
to keep the collector configuration similar to the agent one. Username
and passwords are only collected from environment variables anyway