add a model of the cache mutex design (#275)
The cache has a subscription mechanism: if we request a key from the cache and it is not present, we will do the work to calculate it then put it in the cache. But if there's already a task calculating it, we will instead subscribe to it and get the value once it is generated.

This commit introduces an optimization for this process, and a TLA+ model proving it will work. When we're done calculating the value, the previous code was doing:
-  lock the cache and the wait_map (where we subscribe for the value)
- put the value in the cache
- remove the wait_map for this key
- send the value to all subscriptions
- unlock everything

Now we do:
-  lock the cache and the wait_map (where we subscribe for the value)
- put the value in the cache
- remove the wait_map for this key
- unlock everything
- send the value to all subscriptions

This makes sure that while we are sending the value to all subscribers (potentially costly operation if we are cloning large structures), we are not blocking other queries to the cache