address review comments and fix new telemetry issue
I added the comment as per bryn's review. When testing the comment
change (you can't be too careful), I discovered that the tests kept
hanging in the config_by_file_watching() test.

You can imagine that puzzled me. Anyway, some amount of debugging later
I figured out that I had set APOLLO_KEY and APOLLO_GRAPH_REF in my
environment. This means that when the test tried to initialize, it
tried to create apollo studio reporting. This meant that
when it tried to re-initialize my agent, the "SINGLE THREADED" test
tokio environment deadlocked as per the issue in the opentelemetry
implementation.

I've done some digging and concluded that I need the multi-threaded
runtime to avoid this deadlock. That means I can't use the test_log
macro, because there is no way to specify that the tokio_test
environment needs to be multi-threaded.

I've updated all the tests which I think might be problematic in this
way.