fix: set `operation_name` tracing attribute to empty on unnamed, single-op requests
When GraphQL operation names are not nececessary to execute an operation
(i.e., when there is only a single operation in a GraphQL document) and the
GraphQL operation is _not_ named (i.e., it is anonymous), the
`operation_name` attribute on the trace spans that are associated with the
request will no longer contain a single hyphen character (`-`) but will
instead be an empty string.  This matches the way that these operations are
represented during the GraphQL operation's life-cycle as well.

Fixes #481