further remove the `-` as a pattern for unnamed operation pattern.
As noted in [comment][1], the previous [commit][2] I made wasn't sufficient to
avoid exposing the (meant to be internal) implementation detail of the `-`
(hyphen) character being used as a representation for "anonymous operation"
since all my previous comment did was remove it from the tracing span
attribute where it is later extracted from by Spaceport.

This commit takes things a step further, and updates the `stats_report_key`
method to take a required first parameter of a string.  That string _may_ be
empty, but empty merely matches the expectation of unnamed operation.  Keep
in mind that operations only MUST be named when there are multiple
operations in the document.  In the case of multiple operations, the
out-of-band `operationName` parameter must be passed to the request (either
via POST-body or GET query parameters).

[1]: https://github.com/apollographql/router/pull/525#discussion_r812124165
[2]: https://github.com/apollographql/router/pull/525/commits/adb7bdde9e741a8f7506cc5d77454f20bb824cec