handle the response validation inside response formatting
this makes more sense to validate while we are already exploring the
response tree.
It is also necessary to have the selection sets availalble when
validating, otherwise we get errors on nonNull fields even if they are
not present

This introduces a limited kind of query validation, since we need to
check that some elements of the query match the schema. That validation
returns a None on failure instead of a meaningful error message. We
should wait for apollo-rs semantic analysis to provide good error
messages instead (and delete all of that code)