improvement: Introduce Checkpoint and Step mechanism.
Followup to our recent #537 investigations, we noticed Service layers were missing a straightforward way to let the service orchestrator know whether it shall make progress or bail out.

This Commit introduces `CheckpointLayer` and `CheckpointService` which take a `fn(Request) -> Step<Request, Response>` callback.

`Step` is an enum that contains two variants.
- `Step::Forward(Request)` notifies the CheckpointLayer it should resume calling downstream services.
- `Step::Return(Response)` notifies the CheckpointLayer it should "early return" the response, and not call downstream services.

Iterating on this concept we can now use this to create Layers:
- `EnsureQueryPresence` will Return with a `BAD_REQUEST` if no query is available after the `Automated Persisted Queries` plugin has been called.
- `ForbidHttpGetMutations` will Return with a `METHOD_NOT_ALLOWED` if a mutation was sent over a GET request.
- `APQ` will Return with a `OK` and a graphql error `PERSISTED_QUERY_NOT_FOUND` if the query couldn't be found in the cache.

This refactoring also outlined a couple of bugs that were fixed:
- Backpressure: warp http server returns `SERVICE_UNAVAILABLE` if the `RouterService` isn't ready.
- Error handling: Returning BoxError in a tower service will result in a graphql error and a `INTERNAL_SERVER_ERROR` status.