if we already know the type of a selection, we might not get __typename
in some cases, an object might not contain a typename. Example:

schema:
type Query {
  get: Thing
}

type Thing {
  name: String
}

query {
  get {
    ... on Thing {
      name
    }
  }
}

Here the operation will directly be handled by the subgraph, and it
might not return a __typename field, since we already know which type
will be returned (while for a union or interface we will get a
__typename).

Now when parsing the query we need to store the "current type", which is
known because the Selection::from_ast function is called from the parent
selection creation. If that type and the fragment's type condition
match, we won't need to check it.
There's an additional subtlety here: fragments used as fragment spread
might be defined in the query, not the schema, and might be later in the
query. Since we're lazily going through the query's AST, it makes sense
here to store the current type, then when applying the fragment spread,
look up the actual fragment and check the expected type there