support duplicate selections
we can have multiple sub selections of a same fields coming from
multiple path (by merging operations, or using fragments), so we cannot
remove the field value to avoid a copy. Instead we should keep it in
place and copy whatever we need. This will result in a lot of copies,
but hopefully the ByteString type will reduce their impact