This switches the context RwLock<Obejct> to a DashMap<String, Value>
DashMap is a concurrent hashmap that can be shared between threads.

The advantages are:

Can be used easily in a non-async context.
Provides an alter aka. merge method for performing guaranteed serialized mutations.
Also good for use in async code!
Avoids the potential anti-pattern of users locking context over an await and thus causing all requests to be serialized:
async fn foo() {
  let l = context.extenstions.write().await;
  let v = l.get("foo");
  do_something(v).await; //Oh no!
  l.insert("bar", "bif");
}
As the docs for https://docs.rs/tokio/latest/tokio/sync/struct.Mutex.html indicate. We should avoid async locks unless we know we need to hold a lock across an await boundary. However in our case holding the lock over an await is always wrong and will cause severe perf issues.

An example is demonstrated in the hello plugin, most of the horribleness is due to the use of Value and HeaderValue.
Users are not given direct access to the dashmap so that we can provide extra logic to convert to and from Value. We use Value to ensure that data stored in context can be send to scripting engines.