support duplicate selections (#574)
* support duplicate selections

we can have multiple sub selections of a same fields coming from
multiple path (by merging operations, or using fragments), so we cannot
remove the field value to avoid a copy. Instead we should keep it in
place and copy whatever we need. This will result in a lot of copies,
but hopefully the ByteString type will reduce their impact

* make the input object mutable to avoid copies

* the fragment's known type should use the inner type name

otherwise the fragment's type condition might not match a List or
NonNull

* a null value should not replace a real one when merging json

* pass the output value as mutable argument to format_value

we take value from the input and replace it with null, so if a non
nullable value is asked in multiple selections, we must check if it is
not already there in the output value. But format_value was creating
emlpty objects instead of modifying the main output object, so now we
need to pass it as argument too

Example of failing query:

query {
  me {
    id
  }
  me {
    id
    name
  }
}

after the first "me" operation, "id" was removed from the input object,
but the created output object was not passed to the second "me"
operation, so it could not find id in input nor output