working solution based on reloadable layers
However: due to limitations in FmtSubscriber, we can no longer choose
the type of subscription we would like to have at run time. Rather, this
must be defined statically and we have to choose a single reporting
format and stick to it.

For now, I'm not going with the JSON format.