Fix merge with main so that things compile
I don't know why, but if we use the ServiceBuilder::buffer() method for
certain services, then the code won't compile.

Luckily, I can substitute an explicit Buffer::new() in those places,
which has the same result.

Theory: I think the breakage is caused by the fact that when we use
Buffer, the type must express Buffer<blah...> rather than
BoxService<Bluffer<blahhh>> and that's what causes the issue. It's
really hard to debug though because the compiler doesn't tell you that
but complains about not being able to clone or services not meeting
trait bounds or lifetimes of '0 or ... :)