kick tracing back into life on top of the simplified tracing
Some cosmetic changes to parameter ordering and some changes to
eliminate explicitly (but redundantly) naming a parent.

Also: favour the instrument macro where possible and (most importantly)
set the "plan" probe back to info level so that we can see where we are
spending time when we run a query.

Also: set a default set of trace levels for crates which excludes random
traces from crates that we aren't interested in capturing (I'm looking
at you reqwest).

These changes give me rational behaviour in jaeger in local testing. I'm
not claiming it's perfect, there's stuff we can still improve, but it
should be a big improvement on what we had before.