Re-introduces support for JSON logging
The new subscriber, RouterSubscriber, is uses as our main subscriber.
It's a simple enum which wraps two "flavours" of FmtSubscriber.

All operations are simple delegations to the wrapped FmtSubscriber.

Now that we have our own subscriber, if we wish to change the
implementation in future or add additional configuration options, then
it should be fairly simple.

fixes: #589