handle errors on accept() calls
previously we were unwrapping the result inside the connection's spawned
task, while we can be a bit smarter about errors.
For most of them, we can just ignore the error and continue to the next
connection.
Errors returned when we reach the max file descriptor limit need special
handling, because they're frequently caused by misconfiguration, and
logging a message each time the error is returned will quickly be
overwhelming. So we show an error message once, and then we wait for a
minute before printing another error