docs: Start to introduce information architecture for Router Preview docs.
This brings in some docs from some other places and re-jiggers some of the
existing Managed Federation stuff.

Note that for the moment, it doesn't actually keep the documentation for
Spaceport public as it needs a bit more massaging to make that crisp.
(Someone can definitely do that, but it just needs to move from
`spaceport-configuration-not-currently-linked-in-gatsby-config` into the
`managed-federation/setup` file's workflow.