Fix trace propagation across processes
The "reload" layer which we added as part of the work to simplify
logging and tracing and introduce a tracing plugin, breaks the delicate
chain of downcasting which tracing and opentelemtry tracing expect to
exist.

This fix restores that chain by copying the Reload layer implementation
from tracing-subscriber into our codebase and adding support for
delegated downcasting.

I've also modified our RouterSubscriber to perform delegated downcasting
and recognise when it is being asked to downcast to itself.

I've also fixed a small bug which would have meant that tracing was not
disabled when configuration was set to turn off tracing.