fix: inline configuration subschemas.
Configuration subschemas used to be a reference in the generated yml, which would incur naming conflicts (two configurations with the same name would expand to the same ref thus overriding each other).

While this has been fixed in the code, the test had not.

This commit fixes this and updates the generated snapshots.