docs: configuration file consistency
This uses `router.yaml` in all locations, rather than other variations.
This commit also introduces a TODO for something that came to mind while I
was reading the docs (unexplained context for the configuration file itself)
and also renames the `open-telemetry.mdx` to `opentelemetry.mdx` for
consistency with other locations on our site (and arguably, the name of the
product anyhow).