implement skip and include (#626)
* implement skip directive post processing

we cannot interpret the skip directive in query planning right now, but
we can filter fields after the response was built. This introduces
variables and default values parsing, as those directives can be driven
by variables. The current implementation is a bit wasteful in the way it
considers variables (by copying all default values and variables in a
newly allocated object), but that should be fine if the number of
variables is small. Otherwise, we should have a kind of "formatting
context" containing rhe query, its variables, etc instead of passing
everything as argument to format and apply_selection_set.

* implement the include directive in post processing