improve async cancellation handling in the router
If a client cancelled a request this could cause the router to cancel a
deduplicaton task in the middle of processing. If this happened, the
wait_map would be left in an invalid state and further queries with the
same request would be left waiting for a broadcast which can never
arrive.

This change fixes the dedup handling so that we use a Weak reference to
the broadcast channel which gives us the ability to detect cancellation
events. If a broadcaster has been cancelled, the Weak reference will no
longer be upgradeable and so we know we can clean up and try again.

The problem hasn't manifested in the cache layer yet, as far as we know,
but I've added the fix into the cache layer as well.

There may be other components in the router which are subject to this
problem. We are not aware of any at this time, but we'll need to remain
vigilant against async cancellation issues of this type.