Update docker file to be more usable
Review comments noted we shouldn't specify "-s" flag. This version only
has one default parameter (for a configuration) because the default
listen address (127.0.0.1) is almost always going to be un-helpful in a
container. Minimally, we should have a config which specifies (0.0.0.0)
as the listen address.

I've split CMD into an entrypoint and a default command which just
specifies a config which must be supplied. If the user wishes to
override CMD, they can easily from the command line. If they wish to
supply ENV vars or a new entrypoint (e.g.: for debug shell) then that is
also easily done from command line via --entrypoint parameter.

I've added an author label that references the repo as well.